Fix _swift_makeAnyHashableUpcastingToHashableBaseType to use the witness of the just unboxed typed of a SwiftValue
This got exposed as part of the change to use witnesses as part of the
metadata cache key.

The  Swift._ConcreteHashableBox<MinimalHashableValue> metadata created as
part of:

 AnyHashable(MinimalHashableValue(1))

and
 AnyHashable(_bridgeAnythingToObjectiveC(MinimalHashableValue(1)) as!  NSObject)

Would not be unique to the same metadata (because the witness of
NSObject that we queried before that change was not equal to the witness
of MinimalHashableValue.

rdar://24958043