Parsable interface printer for opaque return types.
Set up a mode for the ASTPrinter where opaque type decls are printed as named, sibling decls
to the decl whose return type they represent, using a special decl syntax. Print the
original decl with an attribute so that we can relate it back to the named opaque decl when
parsing the interface.