[TBDGen] Add async function pointers.
In order to call async functions, instances of the AsyncFunctionPointer
struct must be used.  If those functions are exported from a module, the
AsyncFunctionPointer by means of which the function is to be called must
be exported as well.

For now, the symbol is exported by manually appending the relevant
suffix to the mangled name of the function.